+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Radio button to delivery filter.
+
 ## [3.136.1] - 2025-02-26
 
 ### Removed