+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+-  `showShippingFacet` as `false` by default.
+
 ## [3.137.0] - 2025-06-26
 
 ### Added